made things DRY
