+++ src/components/ActionCell/index.js
@@ -1,0 +1,3 @@
+import ActionCell from './ActionCell';
+
+export default ActionCell;
