+++ src/components/EditCell/index.js
@@ -1,0 +1,3 @@
+import EditCell from './EditCell';
+
+export default EditCell;
