+++ src/routes/list/members/DataTable.js
@@ -21,6 +21,8 @@ import {
 import data from './users';
 import DrawerView from './DrawerView';
 import SearchBar from '../../../components/SearchBar';
+import EditCell from '../../../components/EditCell';
+import ActionCell from '../../../components/ActionCell';
 
 const { Column, HeaderCell, Cell } = Table;
 const { getHeight } = DOMHelper;
@@ -28,40 +30,6 @@ const { getHeight } = DOMHelper;
 type Props = {};
 type State = {
   showDrawer: boolean
-};
-
-export const EditCell = ({ rowData, dataKey, onChange, ...props }) => {
-  const editing = rowData.status === 'EDIT';
-  return (
-    <Cell {...props} className={editing ? 'table-content-editing' : ''}>
-      {editing ? (
-        <input
-          className="rs-input"
-          defaultValue={rowData[dataKey]}
-          onChange={event => {
-            onChange && onChange(rowData.id, dataKey, event.target.value);
-          }}
-        />
-      ) : (
-          <span className="table-content-edit-span">{rowData[dataKey]}</span>
-        )}
-    </Cell>
-  );
-};
-
-const ActionCell = ({ rowData, dataKey, onClick, ...props }) => {
-  return (
-    <Cell {...props} style={{ padding: '6px 0' }}>
-      <Button
-        appearance="link"
-        onClick={() => {
-          onClick && onClick(rowData.id);
-        }}
-      >
-        {rowData.status === 'EDIT' ? 'Save' : 'Edit'}
-      </Button>
-    </Cell>
-  );
 };
 
 class DataList extends React.Component<Props, State> {
