+++ src/routes/list/payroll/Payroll.js
@@ -20,6 +20,8 @@ import {
 import data from '../members/users';
 import DrawerView from '../members/DrawerView';
 import SearchBar from '../../../components/SearchBar';
+import EditCell from '../../../components/EditCell';
+import ActionCell from '../../../components/ActionCell';
 
 const { Column, HeaderCell, Cell } = Table;
 const { getHeight } = DOMHelper;
@@ -49,39 +51,6 @@ const FooterSummary = ({ title, summary }) => (
     </div>
 );
 
-export const EditCell = ({ rowData, dataKey, onChange, ...props }) => {
-    const editing = rowData.status === 'EDIT';
-    return (
-        <Cell {...props} className={editing ? 'table-content-editing' : ''}>
-            {editing ? (
-                <input
-                    className="rs-input"
-                    defaultValue={rowData[dataKey]}
-                    onChange={event => {
-                        onChange && onChange(rowData.id, dataKey, event.target.value);
-                    }}
-                />
-            ) : (
-                    <span className="table-content-edit-span">{rowData[dataKey]}</span>
-                )}
-        </Cell>
-    );
-};
-
-const ActionCell = ({ rowData, dataKey, onClick, ...props }) => {
-    return (
-        <Cell {...props} style={{ padding: '6px 0' }}>
-            <Button
-                appearance="link"
-                onClick={() => {
-                    onClick && onClick(rowData.id);
-                }}
-            >
-                {rowData.status === 'EDIT' ? 'Save' : 'Edit'}
-            </Button>
-        </Cell>
-    );
-};
 
 class Payroll extends React.Component<Props, State> {
     constructor() {
